+++ NEWS.md
@@ -1,6 +1,7 @@
 # macs2r 0.0.1
 
 * Added a `NEWS.md` file to track changes to the package.
-* Added function `import.narrowPeaks`
-* Added Travis CI
-* Added codecov
+* Added function `import.narrowPeaks`.
+* Added Travis CI.
+* Added `covr::package_coverage`.
+* Added `testthat`.